pyproject.toml
[tool.poetry]
name = "iwf-python-samples"
version = "0.1.0"
description = "iWF Python Samples"
authors = ["Quanzheng Long <qlong@indeed.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "signup" }]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.3.3"
iwf-python-sdk = "^0.10.0"
attrs = "^22.2.0"
openai = "^1.66.2"
openai-agents = "^0.0.4"
pydantic = "^2.10.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
mypy = "^1.5.1"
pytest-cov = "^4.1.0"
pytest-env = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
